+++ src/components/Banners/Banner/index.tsx
@@ -1,0 +1,1 @@
+export * from './Banner';
