--- src/components/RenProject/index.tsx
@@ -1,2 +1,0 @@
-export * from './RenBanner';
-export { useLocalStorageBoolean } from '../../helpers/useLocalStorageBoolean';
