+++ src/features/home/home.tsx
@@ -7,8 +7,8 @@ import { selectIsVaultListAvailable } from '../data/selectors/data-loader';
 import { styles } from './styles';
 import { Vaults } from './components/Vaults';
 import { useAppSelector } from '../../store';
-import { AnnouncementBanner } from '../../components/AnnouncementBanner';
-import { RenBannerHome } from '../../components/RenProject';
+import { AnnouncementBanner } from '../../components/Banners/AnnouncementBanner';
+import { RenBannerHome } from '../../components/Banners/RenBanner';
 
 const useStyles = makeStyles(styles);
